refactor(cactus-form): appease the Sonarqube gods
This one was already fixed, but in a way that made me cringe; I'd rather
keep the filename what it was before and change the export.

CACTUS-850
